requirements.txt
# Core streaming and data processing
confluent-kafka==2.12.0
pyspark==3.3.0
apache-airflow==2.9.3

# Storage and search
elasticsearch==8.11.0
minio==7.2.0
hdfs==2.7.3

# Data manipulation and utilities
pandas==2.1.4
numpy==1.26.4
pyyaml==6.0.1

# API clients (optional integrations)
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.2.0

# Airflow providers
apache-airflow-providers-apache-spark==5.0.0
apache-airflow-providers-apache-kafka==1.3.0
apache-airflow-providers-elasticsearch==5.1.0